feat: add details test
