--- src/tests/mockData.ts
@@ -1,12 +1,0 @@
-import { ShowsProps } from '../modules/interfaces';
-
-export const mockCard: ShowsProps = {
-  id: '1',
-  titleOriginal: 'Test Card',
-  totalSeasons: '1',
-  rating: 0,
-};
-
-
-
-
